requirements.txt
# ============================================================
# Python Dependencies for Azure OpenAI Session
# ============================================================
# Install with: pip install -r requirements.txt
# ============================================================

# OpenAI Python SDK - Works with both OpenAI and Azure OpenAI
openai>=1.0.0

# Load environment variables from .env file
python-dotenv>=1.0.0

# Azure AD authentication (token-based, no API key needed)
azure-identity>=1.15.0
